requirements.txt
# requirements.txt
# Kurmak için: pip install -r requirements.txt

python-dotenv>=1.0.0
selenium>=4.15.0
webdriver-manager>=4.0.1
pandas>=2.1.0
openai>=1.3.0
tenacity>=8.2.0
tqdm>=4.66.0